新增UI预制体、新增脚本、优化逻辑
制作治疗文本、护盾文本预制体；新增对应预制体脚本；
移动热更新文件夹战斗UI模块的位置；
新增治疗、护盾方法入口；
伤害链新增护盾处理器节点；
新增UI相关的配置信息；
完善部分枚举脚本；
新增治疗、护盾相关事件；
移除HpChangedEvent的DeltaHp ；
封装怪物血量UI管理逻辑；
LayerUtility封装递归设置层级方法；
LayerGeter新增预定义的层级数值，新增方法；
新增BattleCameraManager封装相机创建逻辑；替换原有逻辑；
BattleEventScheduler新增事件监听，监听角色切换技能选择的逻辑；处理摄像机视角；

+++ Assets/Scripts/Game/Battle/PointInfo.cs
@@ -18,9 +18,9 @@ namespace Game.Battle
         /// <summary>
         /// 怪物中心点x值
         /// </summary>
-        public byte MonsterCenterX { get; }
+        public float MonsterCenterX { get; }
 
-        public PointInfo(Transform point, IBattleEntityObject battleEntity, byte monsterCenterX)
+        public PointInfo(Transform point, IBattleEntityObject battleEntity, float monsterCenterX)
         {
             Point = point;
             BattleEntity = battleEntity;
